python/requirements.txt
requests==2.18.4
pbkdf2==1.3
python-dotenv==0.14.0
pycryptodome==3.9.8